Refactor TokenEncryptionHandler to test and concrete version
